+++ HISTORY.rst
@@ -10,6 +10,7 @@ v0.28 (unreleased)
 * allow dumping dataclasses to JSON, #575 by @samuelcolvin and @DanielOberg
 * ORM mode, #562 by @samuelcolvin
 * fix ``pydantic.compiled`` on ipython, #573 by @dmontagu and @samuelcolvin
+* add `StrictBool` type, #579 by @cazgp
 
 v0.27 (2019-05-30)
 ..................